Add Optional Plug Module for Using with Plug
`Unleash.Plug` will attempt to construct a context for `Unleash` and
place it in `Plug.Conn`'s assigns. It also comes with a helper
`enabled?/3` that takes a `Plug.Conn` to check if a given feature is
enabled.

`Unleash.Plug` also supports `PhoenixGon` for pushing feature flags to
the client.